% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitignore.R
\name{add_gitignore}
\alias{add_gitignore}
\title{Add \code{.gitignore} Templates}
\usage{
add_gitignore(includes, path = getwd(), ...)
}
\arguments{
\item{includes}{(character) The \code{.gitignore} templates to include}

\item{path}{(character) The path to the \code{.gitignore} file. Default is \code{getwd()}}

\item{...}{Additional arguments}
}
\value{
The \code{.gitignore} file with the templates included
}
\description{
Add \code{.gitignore} Templates
}
