% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdown.R
\name{build_pkgdown_with_reports}
\alias{build_pkgdown_with_reports}
\title{No Clocks \code{pkgdown}}
\usage{
build_pkgdown_with_reports(
  pkg = getwd(),
  pkgdown_path = "public",
  assets_path = "pkgdown/assets",
  reports = c("coverage", "testdown", "gitdown"),
  git_branch_ref = "main",
  overwrite_assets = TRUE
)
}
\arguments{
\item{pkg}{(character) Path to the package in development}

\item{pkgdown_path}{(character) Relative path inside the package to store
the final \code{pkgdown} site}

\item{assets_path}{(character) Relative path within the package to store
the \code{pkgdown} assets}

\item{reports}{(character) Vector of reports to be produced. Must be a subset
of \code{c("testdown","gitdown", "coverage")}. The default is
\code{c("coverage", "testdown", "gitdown")}.}

\item{git_branch_ref}{(character) Git branch to use. `main' by default}

\item{overwrite_assets}{(logical) Whether the assets directory should be
overwritten. Default is \code{TRUE}.}
}
\value{
None Generate a pkgdown with test and coverage reports
}
\description{
Build a \code{\link[pkgdown:pkgdown-package]{pkgdown::pkgdown()}} site with additional options to include
extra reports.

The reports that can be included are:
\itemize{
\item Test Results Report via \code{\link[testdown:test_down]{testdown::test_down()}}
\item Test Coverage Results via \code{\link[covrpage:covrpage]{covrpage::covrpage()}}
\item Git Reports via \code{\link[gitdown:git_down]{gitdown::git_down()}}
}
}
\examples{
\dontrun{
# build_pkgdown_with_reports(
#    pkg = ".",
#    pkgdown_path = "public",
#    assets_path = "pkgdown/assets",
#    reports = c("testdown","coverage")
}
}
