% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfg_init.R
\name{cfg_d_init}
\alias{cfg_d_init}
\title{Configuration Sub-Directory Initialization}
\usage{
cfg_d_init(
  path = here::here("inst/config"),
  configs = list(example = list(default = list(NULL))),
  merge = TRUE,
  ignore = TRUE,
  overwrite = FALSE,
  templates = FALSE,
  include_encrypted = FALSE,
  ...
)
}
\arguments{
\item{path}{Path to the configuration directory. Defaults to \code{inst/config}.}

\item{configs}{A list of configuration lists for different environments or purposes.
The list should contain named lists of configuration settings where the
names represent what will be used for the name of the configuration file.
See details for more information and examples.}

\item{merge}{Merge the configuration files from the \verb{config.d/} directory into
the \code{config.yml} file? Defaults to \code{TRUE}.}

\item{ignore}{Add a \code{.gitignore} to the \verb{config.d/} directory to ignore the
generated configuration files? Default is \code{TRUE}.}

\item{overwrite}{Overwrite the existing configuration files. Defaults to \code{FALSE}.}

\item{templates}{Create configuration template files for each configuration file?
Defaults to \code{FALSE}.}

\item{include_encrypted}{Create encrypted configuration files for each configuration file?
Defaults to \code{FALSE}.}

\item{...}{Additional arguments passed to other functions.}
}
\value{
Invisibly returns the merged configuration settings.
}
\description{
This function initializes the configuration sub-directory setup by creating
a sub-directory: \verb{config.d/} in the root \verb{config/} directory to house
configuration files for different environments or purposes that can be
merged together.
}
\seealso{
\code{\link[=cfg_init]{cfg_init()}} and \code{\link[config:merge]{config::merge()}}
}
