% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfg_encrypt.R
\name{cfg_encrypt}
\alias{cfg_encrypt}
\alias{encrypt_cfg_file}
\alias{decrypt_cfg_file}
\title{Configuration File Encryption/Decryption}
\usage{
encrypt_cfg_file(
  cfg_file = Sys.getenv("R_CONFIG_FILE", "inst/config/config.yml"),
  key = "NOCLOCKS_ENCRYPTION_KEY",
  overwrite = FALSE,
  ...
)

decrypt_cfg_file(
  cfg_file = Sys.getenv("R_CONFIG_FILE", "inst/config/config.yml"),
  cfg_file_encrypted = .get_encrypted_cfg_file(cfg_file),
  key = "NOCLOCKS_ENCRYPTION_KEY"
)
}
\arguments{
\item{cfg_file}{Path to the configuration file. Defaults to the \code{R_CONFIG_FILE}
environment variable or \code{inst/config/config.yml}.}

\item{key}{The name of the environment variable that contains the encryption key.
Defaults to \code{NOCLOCKS_ENCRYPTION_KEY}.}

\item{overwrite}{Logical. Overwrite the existing encrypted file. Defaults to \code{FALSE}.
Used in \code{encrypt_cfg_file()} only.}

\item{...}{Additional arguments passed to other functions.}

\item{cfg_file_encrypted}{Path to the encrypted configuration file. Defaults to the
\code{cfg_file} with the extension \code{.encrypted.yml}. (Used in \code{decrypt_cfg_file()} only).}
}
\value{
\itemize{
\item \code{encrypt_cfg_file()} returns invisible \code{0}.
\item \code{decrypt_cfg_file()} returns invisible \code{config::get()}.
}
}
\description{
\itemize{
\item \code{encrypt_cfg_file()} encrypts the configuration file using the encryption key.
\item \code{decrypt_cfg_file()} decrypts the configuration file using the encryption key.
}
}
\details{
The \code{encrypt_cfg_file()} function encrypts the configuration file using the encryption key.
The \code{decrypt_cfg_file()} function decrypts the configuration file using the encryption key.
}
\seealso{
\code{\link[=cfg_init]{cfg_init()}} \code{\link[=cfg_hooks_init]{cfg_hooks_init()}}
}
