% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfg_hooks.R
\name{cfg_hooks_init}
\alias{cfg_hooks_init}
\title{Initialize Configuration Encryption Git Hooks}
\usage{
cfg_hooks_init(
  cfg_file = Sys.getenv("R_CONFIG_FILE", here::here("inst/config/config.yml")),
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{cfg_file}{Path to config.yml}

\item{overwrite}{Overwrite existing hooks}

\item{...}{Additional arguments passed to other functions}
}
\description{
Initialize Configuration Encryption Git Hooks
}
