% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfg_init.R
\name{cfg_init}
\alias{cfg_init}
\title{Initialize Configuration Setup}
\usage{
cfg_init(
  path = here::here("inst/config"),
  cfg = list(default = list(NULL)),
  cfg_file = "config.yml",
  cfg_file_encrypted = "config.encrypted.yml",
  cfg_file_template = "config.template.yml",
  encryption_key_name = "NOCLOCKS_ENCRYPTION_KEY",
  overwrite = FALSE,
  symlink = TRUE,
  ...
)
}
\arguments{
\item{path}{Path to the configuration directory. Defaults to \code{inst/config}.}

\item{cfg}{A list of configuration settings. Defaults to an empty list.}

\item{cfg_file}{Path to the configuration file. Defaults to \code{config.yml}.}

\item{cfg_file_encrypted}{Path to the encrypted configuration file. Defaults to \code{config.encrypted.yml}.}

\item{cfg_file_template}{Path to the configuration template file. Defaults to \code{config.template.yml}.}

\item{encryption_key_name}{The name of the environment variable that contains the encryption key.
Defaults to \code{NOCLOCKS_ENCRYPTION_KEY}.}

\item{overwrite}{Overwrite the existing configuration files. Defaults to \code{FALSE}.}

\item{symlink}{Create a symlink to the configuration file. Defaults to \code{TRUE}.}

\item{...}{Additional arguments passed to other functions.}
}
\value{
Invisibly returns the configuration settings.
}
\description{
This function initializes the configuration setup by creating the configuration file,
encrypted configuration file, and configuration template file.

It also sets the \code{R_CONFIG_FILE} environment variable to the path of the configuration file.
}
\details{
The \code{cfg_init()} function creates the following files:
\itemize{
\item \code{config.yml}: The configuration file containing the configuration settings.
\item \code{config.encrypted.yml}: The encrypted configuration file.
\item \code{config.template.yml}: The configuration template file.
\item \code{.gitignore}: The gitignore file to ignore the configuration files.
\item \code{README.md}: The README file with instructions on how to use the configuration files.
}
}
\seealso{
\code{\link[config:get]{config::get()}}
}
