% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theming.R
\name{color_palette}
\alias{color_palette}
\title{No Clocks Color Palette}
\usage{
color_palette()
}
\value{
A named vector of colors
}
\description{
The No Clocks color palette is a set of colors that are used in various
No Clocks, LLC projects. The palette is designed to be visually appealing
and accessible.
}
\examples{

}
\seealso{
\url{No Clocks Colors}

\url{No Clocks Brand Guidelines - Colors}
}
