% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfg_template.R
\name{create_cfg_template}
\alias{create_cfg_template}
\title{Generate a \code{config.template.yml} from a \code{config.yml}}
\usage{
create_cfg_template(input_file, output_file = NULL)
}
\arguments{
\item{input_file}{The path to the input \code{config.yml} file.}

\item{output_file}{The path where the generated \code{config.template.yml} file should be saved.}
}
\value{
None. The function outputs the template file to the specified location.
}
\description{
This function takes a \code{config.yml} file as input and generates a \code{config.template.yml}
file by replacing values with specific placeholders like \code{"<APP_NAME>"}, \code{"<API_KEY>"}, etc.,
and ensuring that the placeholders are correctly quoted in the output YAML.
}
\examples{
generate_config_template("inst/config/config.yml", "inst/config/config.template.yml")
}
