% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{.noclocks_colors}
\alias{.noclocks_colors}
\title{No Clocks Color Palettes}
\format{
An object of class \code{list} of length 11.
}
\usage{
.noclocks_colors
}
\description{
Collection of color palettes for use in No Clocks, LLC projects.
}
\keyword{datasets}
