% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onLoad.R
\name{.onAttach}
\alias{.onAttach}
\title{Startup Functions}
\usage{
.onAttach(libname = find.package("noclocksR"), pkgname = "noclocksR")
}
\description{
These functions are run when the package is loaded or attached.
}
