% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brandfetch.R
\name{download_brand_logos}
\alias{download_brand_logos}
\title{Download Brand Logos}
\usage{
download_brand_logos(brand, path = "inst/extdata/brand", ...)
}
\arguments{
\item{brand}{Brand}

\item{path}{Path}

\item{...}{...}
}
\value{
Invisible
}
\description{
Download Brand Logos
}
