% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brandfetch.R
\name{download_logo}
\alias{download_logo}
\title{Download Brand Logo File}
\usage{
download_logo(
  src,
  file,
  name,
  path = "inst/extdata/brand",
  type = c("icon", "logo"),
  format = c("png", "svg", "jpeg"),
  height,
  width,
  ...
)
}
\arguments{
\item{src}{The URL of the logo file}

\item{file}{The name of the logo file}

\item{name}{The name of the brand}

\item{path}{The path to save the logo file}

\item{type}{The type of logo (icon or logo)}

\item{format}{The format of the logo (png, svg, jpeg)}

\item{height}{The height of the logo}

\item{width}{The width of the logo}

\item{...}{Additional arguments}
}
\value{
Invisible
}
\description{
This function downloads a brand logo file from a URL to the specified path.
}
