% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{extract_pdf_content}
\alias{extract_pdf_content}
\title{Extract PDF Content}
\usage{
extract_pdf_content(path, ...)
}
\arguments{
\item{path}{(Required) The path to the PDF file.}

\item{...}{Additional arguments}
}
\value{
A character vector containing the text content of the PDF.
}
\description{
Extracts the text content from a PDF file.
}
