% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brandfetch.R, R/utils_branding.R
\name{fetch_brand}
\alias{fetch_brand}
\title{Fetch a Brand using the Brandfetch API}
\usage{
fetch_brand(
  domain,
  brandfetch_api_key = Sys.getenv("BRANDFETCH_API_KEY", unset =
    config::get("brandfetch_api_key")),
  ...
)

fetch_brand(
  domain,
  brandfetch_api_key = Sys.getenv("BRANDFETCH_API_KEY", unset =
    config::get("brandfetch_api_key")),
  ...
)
}
\arguments{
\item{domain}{The domain of the brand to fetch}

\item{brandfetch_api_key}{The API key for the Brandfetch API}

\item{...}{Additional arguments}
}
\value{
A tibble with the brand information

A tibble with the brand information
}
\description{
This function fetches a brand using the
\href{https://docs.brandfetch.com/reference/brand-api}{Brandfetch Brand API}.

This function fetches a brand using the
\href{https://docs.brandfetch.com/reference/brand-api}{Brandfetch Brand API}.
}
\examples{

}
