% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_function_definition}
\alias{get_function_definition}
\title{Get Source Code of a Function's Definition}
\usage{
get_function_definition(fun)
}
\arguments{
\item{fun}{An R function}
}
\value{
The source code of the function's definition.
}
\description{
This function retrieves the source code of a function's definition by
simply wrapping \code{attr(fun, "srcref")} to pull the "source reference"
attribute of the function.

This is more useful than simply printing the function definition by
excluding \verb{()} because it provides the actual source code
(and comments, formatting, etc.) of the function definition.
}
\examples{
get_function_definition(print)
}
\seealso{
\code{\link[base:srcfile]{base::srcfile()}}
}
\concept{Utility}
