% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitignore.R
\name{get_gitignore}
\alias{get_gitignore}
\title{\code{.gitignore} Templates}
\usage{
get_gitignore(template, as = c("list", "json"))
}
\arguments{
\item{template}{(character) The name of the \code{.gitignore} template}

\item{as}{(character) The format to return the \code{.gitignore} template. Can be
\code{list} or \code{json}. Default is \code{list}.}
}
\value{
The \code{.gitignore} template as a list or JSON
}
\description{
Get \code{.gitignore} templates from \href{https://www.toptal.com/developers/gitignore}{Toptal}.
}
