% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brandfetch.R
\name{get_logo_file_name}
\alias{get_logo_file_name}
\title{Get Logo File Name}
\usage{
get_logo_file_name(
  brand_name,
  type,
  theme,
  format,
  height = NA,
  width = NA,
  ...
)
}
\arguments{
\item{brand_name}{Brand Name}

\item{type}{Type}

\item{theme}{Theme}

\item{format}{Format}

\item{height}{Height}

\item{width}{Width}

\item{...}{...}
}
\value{
The logo file name
}
\description{
Get Logo File Name
}
