% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_attributes.R
\name{git_attributes}
\alias{git_attributes}
\title{Git Attributes}
\usage{
git_attributes(...)
}
\arguments{
\item{...}{...}
}
\value{
...
}
\description{
...
}
\examples{
if (FALSE) {

  git_attributes()

}
}
