% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_minify.R
\name{minify_css}
\alias{minify_css}
\title{Minify CSS}
\usage{
minify_css(css)
}
\arguments{
\item{css}{A character vector of CSS code}
}
\value{
A character vector of minified CSS code
}
\description{
Minify CSS
}
