% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noclocksr-package.R
\docType{package}
\name{noclocksr-package}
\alias{noclocksr}
\alias{noclocksr-package}
\title{noclocksr: No Clocks Utilities and Helpers}
\description{
Library created for internal development at No Clocks.
}
\author{
\strong{Maintainer}: Jimmy Briggs \email{jimmy.briggs@jimbrig.com} (\href{https://orcid.org/0000-0002-7489-8787}{ORCID})

Authors:
\itemize{
  \item Patrick Howard \email{patrick.howard@noclocks.dev} (\href{https://orcid.org/0000-0000-0000-0000}{ORCID}) [reviewer]
}

Other contributors:
\itemize{
  \item No Clocks, LLC \email{dev@noclocks.dev} [funder, copyright holder]
}

}
\keyword{internal}
