% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{parse_pdf_content}
\alias{parse_pdf_content}
\title{Parse PDF Content}
\usage{
parse_pdf_content(pdf_content, ...)
}
\arguments{
\item{pdf_content}{(Required) The text content extracted from the PDF file.}

\item{...}{Additional arguments}
}
\value{
A list containing parsed details: type, date, company, id, new_file_name.
}
\description{
Parses the extracted text content from a PDF file.
}
