% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_debug_button.R
\name{shiny_debug_button}
\alias{shiny_debug_button}
\alias{shiny_debug_button_ui}
\alias{shiny_debug_button_server}
\title{Shiny Debug Button}
\usage{
shiny_debug_button_ui(
  id = "browser",
  label = "Debug",
  icon = shiny::icon("bug"),
  hide = TRUE,
  keyboard_shortcut = "Ctrl + Alt + D"
)

shiny_debug_button_server(id = "browser")
}
\arguments{
\item{id}{The ID of the button. Default is \code{browser}.}

\item{label}{The label for the button.}

\item{icon}{The icon to use for the button. Default is a bug icon.}

\item{hide}{Whether to hide the button by default. Default is \code{TRUE}.}

\item{keyboard_shortcut}{The keyboard shortcut to toggle the button's visibility.
Default is \code{Ctrl + Alt + D}.}
}
\value{
A UI element containing the hidden button and associated JavaScript.
}
\description{
Shiny Module that creates a special \code{DEBUG} button to toggle the \code{browser()} function in a Shiny application.

\itemize{
\item \code{shiny_debug_button_ui()}: Creates the UI for the Shiny Debug Button.
}

\itemize{
\item \code{shiny_debug_button_server()}: Handles the server-side logic for the Shiny Debug Button.
}
}
\details{
The \code{shiny_debug_button_ui} function creates a hidden button in the Shiny UI that can be toggled to run the \code{browser()} function.
The \code{shiny_debug_button_server} function contains the server-side logic to handle the button's functionality.

This function creates a button that can be used to toggle the \code{browser()} function in a Shiny application.
The button is hidden by default and can be toggled by pressing the keyboard shortcut \code{Ctrl + Alt + D}
(or whatever the user specifies via the \code{keyboard_shortcut} argument).
}
\section{Functions}{

\describe{
\item{\code{\link{shiny_debug_button_ui}}}{Creates the button UI.}
\item{\code{\link{shiny_debug_button_server}}}{Handles server-side logic.}
}
}

\examples{
# launch a shiny app with an interactive debug button
# that can be toggled on and off via keyboard shortcut
# the default keyboard shortcut is Ctrl + Alt + D:

if (interactive()) {

  library(shiny)

  ui <- fluidPage(
    shiny_debug_button_ui(id = "browser_btn", label = "Debug")
  )
  server <- function(input, output, session) {
    shiny_debug_button_server("browser_btn")
  }

  shinyApp(ui, server)

}

}
