% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_resume.R
\name{shiny_resume_body}
\alias{shiny_resume_body}
\title{Shiny Resume Body}
\usage{
shiny_resume_body(...)
}
\arguments{
\item{...}{A series of \code{\link[=shiny_resume_section]{shiny_resume_section()}} elements}
}
\value{
HTML via \code{\link[htmltools:tagList]{htmltools::tagList()}}
}
\description{
Produce the body of the shiny resume page.
}
