% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_resume.R
\name{shiny_resume_navbar}
\alias{shiny_resume_navbar}
\title{Shiny Resume Navigation Bar}
\usage{
shiny_resume_navbar(
  refs = c(a = "A"),
  image = "img/profile.jpg",
  color = "#bd5d38"
)
}
\arguments{
\item{refs}{Named list with names being IDs and elements being text displayed.}

\item{image}{Image to place in navbar}

\item{color}{Color of the sidebar}
}
\value{
HTML via \code{\link[htmltools:tagList]{htmltools::tagList()}}
}
\description{
Produce a navigation bar for a shiny resume
}
