% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_resume.R
\name{shiny_resume_page}
\alias{shiny_resume_page}
\title{Shiny Resume Page}
\usage{
shiny_resume_page(app_title, nav, body)
}
\arguments{
\item{app_title}{Title to use for the app}

\item{nav}{Navbar, constructed with \code{\link[=shiny_resume_navbar]{shiny_resume_navbar()}}}

\item{body}{Body, constructed with \code{\link[=shiny_resume_body]{shiny_resume_body()}} & \code{\link[=shiny_resume_section]{shiny_resume_section()}}}
}
\value{
HTML via \code{\link[htmltools:tagList]{htmltools::tagList()}}
}
\description{
Shiny App with Bootstrap Resume Template
}
