% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_toggl.R
\name{time_tracking}
\alias{time_tracking}
\alias{start_time_tracking}
\alias{stop_time_tracking}
\alias{get_tracked_time}
\title{Toggl Time Tracking}
\usage{
start_time_tracking(
  description = "R Development for GMH Leasing Dashboard",
  tags = c(),
  config = config::get("toggl"),
  ...
)

stop_time_tracking(...)

get_tracked_time(
  start = Sys.time() - lubridate::weeks(1),
  end = Sys.time(),
  ...
)
}
\arguments{
\item{description}{A description of the time entry.
Default is "R Development for GMH Leasing Dashboard" in this project.}

\item{tags}{A character vector of tags to apply to the time entry.
Note that if the project is billable, the "Billable" tag will be added.}

\item{config}{A configuration list for the Toggl project.
By default will retrieve values from the \code{toggl} configuration setup in
the \code{config.yml} for the project.}

\item{...}{Additional arguments to pass to the various \code{togglr} functions.}
}
\value{
\itemize{
\item \code{start_time_tracking()}: The response from the Toggl API for starting time tracking.
\item \code{stop_time_tracking()}: The response from the Toggl API for stopping time tracking.
\item \code{get_tracked_time()}: A data frame of the time entries retrieved from Toggl.
}
}
\description{
Functions for tracking time in the current project's context via Toggl.
\itemize{
\item \code{start_time_tracking()}: Start tracking time in Toggl.
\item \code{stop_time_tracking()}: Stop tracking time in Toggl.
\item \code{get_tracked_time()}: Retrieve tracked time entries from Toggl.
}
}
