% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theming.R
\name{typography}
\alias{typography}
\title{No Clocks Typography (Fonts)}
\usage{
typography()
}
\value{
A named vector of fonts
}
\description{
The No Clocks typography is a set of fonts that are used in various
No Clocks, LLC projects. The typography is designed to be visually appealing
and accessible.
}
\examples{

}
\seealso{
\url{No Clocks Typography}
}
