% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_meta.R
\name{use_noclocks_meta}
\alias{use_noclocks_meta}
\title{Use No Clocks \verb{<meta>}}
\usage{
use_noclocks_meta(
  name = "noclocks",
  version = "0.0.1",
  description = "<meta> and social media cards for No Clocks, LLC",
  url = "https://noclocks.dev",
  theme_color = "#000000",
  robots = "index,follow",
  generator = "R-Shiny",
  subject = "No Clocks, LLC",
  rating = "General",
  referrer = "origin",
  csp = "default-src 'self'",
  image = noclocks_logo(url = TRUE),
  image_alt = "No Clocks, LLC Logo",
  twitter_creator = "@noclocksdev",
  twitter_card_type = "summary_large_image",
  twitter_site = "@noclocksdev",
  ...
)
}
\arguments{
\item{name}{(character) The name of the site}

\item{version}{(character) The version of the site}

\item{description}{(character) The description of the site}

\item{url}{(character) The URL of the site}

\item{theme_color}{(character) The theme color of the site}

\item{robots}{(character) The robots meta tag}

\item{generator}{(character) The generator meta tag}

\item{subject}{(character) The subject meta tag}

\item{rating}{(character) The rating meta tag}

\item{referrer}{(character) The referrer meta tag}

\item{csp}{(character) The content security policy meta tag}

\item{image}{(character) The image URL for the site}

\item{image_alt}{(character) The image alt text for the site}

\item{twitter_creator}{(character) The Twitter creator meta tag}

\item{twitter_card_type}{(character) The Twitter card type meta tag}

\item{twitter_site}{(character) The Twitter site meta tag}

\item{...}{Additional arguments}
}
\value{
HTML via \code{\link[htmltools:builder]{htmltools::tags()}} and \verb{<meta>} tags via
\code{\link[metathis:meta]{metathis::meta()}}
}
\description{
Adds \verb{<meta>} tags and social media cards for No Clocks, LLC.
}
\examples{
if (FALSE) {

  shiny::fluidPage(
    use_noclocks_meta()
  )

}
}
\seealso{
\code{\link[metathis:meta]{metathis::meta()}}
}
