% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_html.R
\name{utils_html}
\alias{utils_html}
\alias{flucol}
\alias{icon_text}
\title{Shiny HTML Utilities}
\usage{
flucol(..., width = 12, offset = 0)

icon_text(icon, text)
}
\arguments{
\item{...}{Elements to include within the columns. These can be any valid
HTML tags or Shiny UI elements.}

\item{width}{Integer for the width to use for the column}

\item{offset}{Integer for the offset to use for the column}

\item{icon}{Either a character string representing a shiny icon or a shiny icon tag
created with \code{\link[shiny:icon]{shiny::icon()}}. If a character string is provided, it will be
converted to a shiny icon tag.}

\item{text}{Character string to display next to the provided icon.}
}
\value{
\itemize{
\item \code{flucol()}: Returns a \code{\link[shiny:fluidPage]{shiny::fluidRow()}} wrapping a \code{\link[shiny:column]{shiny::column()}} containing
the provided UI elements
\item \code{icon_text()}: Returns a \verb{<span>} tag containing the provided icon and text.
}

Returns a \code{\link[shiny:fluidPage]{shiny::fluidRow()}} wrapping a \code{\link[shiny:column]{shiny::column()}} containing
the provided UI elements

Returns a \verb{<span>} tag containing the provided icon and text.
}
\description{
\itemize{
\item \code{flucol()}: Shiny Fluid Row + Column Layout
\item \code{icon_text()}: Icon Text
}

Create a fluid row with multiple columns. This function is a wrapper around
\code{\link[shiny:fluidPage]{shiny::fluidRow()}} and \code{\link[shiny:column]{shiny::column()}}. It allows for the creation of
multiple columns with varying widths and offsets, or just as a shorthand
for \code{shiny::fluidRow(shiny::column(width = 12, ...))}.

Create a \verb{<span>} tag containing an icon and text. Useful for displaying
icons next to text, typically for labels and titles.
}
\details{
These functions provide utility functions for working with HTML and Shiny UI
elements. They are designed to simplify the creation of common UI elements
and layouts.
}
\examples{
flucol(
  htmltools::tags$h1("Hello World!")
)
# provide plain text icon
icon_text("table", "Table") |> htmltools::browsable()

# provide shiny icon tag
icon_text(shiny::icon("table"), "Table") |> htmltools::browsable()
}
\seealso{
\code{\link[shiny:fluidPage]{shiny::fluidRow()}}, \code{\link[shiny:column]{shiny::column()}}, \code{\link[shiny:icon]{shiny::icon()}}
}
