% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfg_validate.R
\name{validate_cfg}
\alias{validate_cfg}
\title{Validate Configuration}
\usage{
validate_cfg(cfg)
}
\arguments{
\item{cfg}{Configuration object (list) to validate}
}
\value{
TRUE if the configuration is valid, otherwise an error is thrown.
}
\description{
A simple function to validate a configuration object that uses the \code{config}
package.
}
\details{
This function verifies that the configuration object is a list and that it
contains at least a 'default' configuration.
}
