% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_dataset.R
\name{document_dataset}
\alias{document_dataset}
\title{Document a Dataset}
\usage{
document_dataset(
  data_obj,
  name = "DATASET_NAME",
  description = "DATASET_DESCRIPTION",
  source = "DATASET_SOURCE",
  col_types = purrr::map_chr(data_obj, typeof),
  col_descs = rep("COLUMN_DESCRIPTION", length(names(data_obj))),
  file = fs::path("R", "data.R"),
  append = TRUE,
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{data_obj}{(Required) The dataset to document. Should be a \code{data.frame}
or \code{tibble}.}

\item{name}{(Optional) The name of the dataset. Default is \code{DATASET_NAME} if
not provided.}

\item{description}{(Optional) A description of the dataset. Default is
\code{DATASET_DESCRIPTION} if not provided.}

\item{source}{(Optional) The source of the dataset. Default is \code{DATASET_SOURCE}
if not provided.}

\item{col_types}{(Optional) A character vector of column types. Default is
the types of the columns in the dataset.}

\item{col_descs}{(Optional) A character vector of column descriptions. Default
is \code{COLUMN_DESCRIPTION} for each column.}

\item{file}{(Optional) The file to write the documentation to. Default is
\code{R/data.R}.}

\item{append}{(Optional) Append to the file if it exists. Default is \code{TRUE}.
If \code{FALSE}, the file will be overwritten, unless \code{overwrite} is \code{TRUE}, then
the function does not write to a file.}

\item{overwrite}{(Optional) Overwrite the file if it exists. Default is \code{FALSE}.}

\item{...}{Additional arguments}
}
\value{
Invisibly returns a character string with the \code{roxygen2}
documentation skeleton for the dataset.
}
\description{
This function generates a \code{roxygen2} skeleton for documenting a dataset.
}
\examples{
data(mtcars)
document_dataset(
  mtcars,
  name = "mtcars",
  description = "Motor Trend Car Road Tests",
  source = "Henderson and Velleman (1981)",
  col_descs = c(
    "Miles/(US) gallon",
    "Number of cylinders",
    "Displacement (cu.in.)",
    "Gross horsepower",
    "Rear axle ratio",
    "Weight (1000 lbs)",
    "1/4 mile time",
    "V/S"
  ),
  file = fs::path_temp("mtcars.R")
)

file.edit(fs::path_temp("mtcars.R"))
}
