% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noclocksR-package.R
\docType{package}
\name{noclocksR-package}
\alias{noclocksR}
\alias{noclocksR-package}
\title{noclocksR: What the Package Does (One Line, Title Case)}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

No Clocks, LLC packaged assets and workflows
}
\author{
\strong{Maintainer}: Jimmy Briggs \email{jimmy.briggs@jimbrig.com} (\href{https://orcid.org/0000-0002-7489-8787}{ORCID})

}
\keyword{internal}
