% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noclocksR-package.R
\docType{package}
\name{noclocksr-package}
\alias{noclocksr}
\alias{noclocksr-package}
\title{noclocksr: Internal Development at No Clocks, LLC}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

No Clocks, LLC packaged assets and workflows
}
\seealso{
Useful links:
\itemize{
  \item \url{https://noclocks.github.io/noclocksr/}
  \item \url{https://docs.noclocks.dev/noclocksr/}
}

}
\author{
\strong{Maintainer}: Jimmy Briggs \email{jimmy.briggs@jimbrig.com} (\href{https://orcid.org/0000-0002-7489-8787}{ORCID})

}
\keyword{internal}
